bot/requirements.txt

python-telegram-bot==22.5
asyncpg==0.31.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
aiosqlite==0.19.0